Extract field guide page hooks
